% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{bh_dd_stages_check_ok}
\alias{bh_dd_stages_check_ok}
\title{bh_dd_stages_check}
\usage{
bh_dd_stages_check_ok(bh_dd_stages = NA)
}
\description{
Check that bh_dd_stages vector values are valid.
}
\keyword{internal}
