% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_k_for_proj_dd.R
\name{build_k_for_proj_dd}
\alias{build_k_for_proj_dd}
\title{build_k_for_proj_dd}
\usage{
build_k_for_proj_dd(habitat_dd_k, HUC_ID, life_histories, life_cycle_params)
}
\description{
Build build_habitat_dd_k object and bh_dd_stages from PopulationModel_Run for Projection_DD
}
\keyword{internal}
