% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{E_est_anadromous}
\alias{E_est_anadromous}
\title{E_est_anadromous}
\usage{
E_est_anadromous(N, dat, spanwing_years)
}
\description{
E_est_anadromous
}
\details{
Egg count - number of egg produced by pop annually for anadromous
}
\keyword{internal}
