% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_PopulationModel.R
\name{get_transition_positions}
\alias{get_transition_positions}
\title{proc utility function}
\usage{
get_transition_positions(matrix_size)
}
\description{
proc utility function
}
\keyword{internal}
