% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_ce_apply.R
\name{pop_model_ce_apply}
\alias{pop_model_ce_apply}
\title{Population Model CE Application to Vital Rates}
\usage{
pop_model_ce_apply(dat = NULL, CE_df = NULL)
}
\arguments{
\item{dat}{Life history vital rates list object.}

\item{CE_df}{Cumulative effects data frame.}

\item{alevin_stage}{Alevin stage name.}

\item{all_juv}{All juvenile stage names.}

\item{fry_stages}{Fry stage names.}

\item{fry_parr_stages}{Fry and parr stage names.}

\item{parr_stages}{Parr stage names.}

\item{subadult_stages}{Subadult stage names.}

\item{adult_stages}{Adult stage names.}
}
\value{
a modified dat vital rates object after applying CE effects.
}
\description{
Utility function to apply cumulative effects to vital rates.
}
