% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_ce_apply_year.R
\name{pop_model_ce_apply_year}
\alias{pop_model_ce_apply_year}
\title{Population Model CE Application to Vital Rates Year-Specific}
\usage{
pop_model_ce_apply_year(
  rates = NULL,
  CE_df = NULL,
  anadromous = FALSE,
  mat_stage = NULL
)
}
\arguments{
\item{rates}{Life history vital rates list object.}

\item{CE_df}{Cumulative effects data frame.}

\item{anadromous}{Anadromous TRUE/FALSE.}

\item{mat_stage}{Named numeric vector of mature life stages.}
}
\value{
a modified dat vital rates object after applying CE effects.
}
\description{
Utility function to apply cumulative effects to vital rates.
}
