% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_dat_clean.R
\name{pop_model_dat_clean}
\alias{pop_model_dat_clean}
\title{pop_model_dat_clean}
\usage{
pop_model_dat_clean(dat, nstage_fill = NULL)
}
\arguments{
\item{dat}{Data frame of life history parameters.}

\item{nstage_fill}{Numeric. Number of stages to fill out to.}
}
\description{
pop_model_dat_clean
}
\details{
Cleans input parameters from csv.
}
