% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_get_spawners.R
\name{pop_model_get_spawners}
\alias{pop_model_get_spawners}
\title{Population Model Get Spawners}
\usage{
pop_model_get_spawners(
  data = NULL,
  life_stage = "spawners",
  life_cycle_params = NULL
)
}
\arguments{
\item{data}{list output from \code{PopulationModel_Run()}.}

\item{life_stage}{Target lifestage. Character string. Can be either \code{spawners} or.}

\item{life_cycle_params}{dataframe. Life cycle parameters.}
}
\value{
dataframe of spawner abundance by location, replicate, and year.
}
\description{
Get spawners from \code{PopulationModel_Run()} output.
}
\details{
Transform verbose output from PopulationModel_Run() into clean dataframe of abundance data for target life stage. This should be run after \code{PopulationModel_Run()} or \code{Projection_DD}.
}
\examples{
\dontrun{
library(CEMPRA)
}

}
