% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_hab_dens_clean.R
\name{pop_model_hab_dens_clean}
\alias{pop_model_hab_dens_clean}
\title{pop_model_hab_dens_clean}
\usage{
pop_model_hab_dens_clean(hab_dens = NA)
}
\arguments{
\item{hab_dens}{A data frame. Habitat density data frame with capacity columns.}
}
\description{
pop_model_hab_dens_clean
}
\details{
Cleans input parameters from csv.
}
