% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_setup_anadromous.R
\name{pop_model_setup_anadromous}
\alias{pop_model_setup_anadromous}
\title{Population Model Setup for Anadromous Species}
\usage{
pop_model_setup_anadromous(life_cycles = NA)
}
\arguments{
\item{life_cycles}{Data frame. Raw life cycles.csv data frame.}
}
\value{
a list object of symbolic objects.
}
\description{
Generate symbolic objects for the population model and
then generate parameter matrices for Pacific salmon.
}
\details{
This is an intermediate setup function to run the population,
but some of its outputs are useful on their own, especially for eigen
analyses from the projection matrix. pop_model_setup_anadromous() is run
before pop_model_matrix_elements() and Projection_DD().

pop_model_setup_anadromous() is a customized version of pop_model_setup() for
anadromous species (Salmon). Follow convention from Davison and Statterthwaite
2016 (Natural Resource Modelling Use of matrix model to predict life history
schedules for semelparous population).

Transition Matrix:
\itemize{
\item Calculates stage-specific survivals and transition rates, survival probabilities, transition probabilities.
\item Initializes the survival/transition rates subject to a stage-specific density-dependence.
\item Checks for excessive compensation ratios.
}

Dataset for life-histories:
\itemize{
\item transition probabilities
\item egg-survival, age-0 survival, and then stage-specific survival probabilities
\item spawning events per year
\item eggs-per-spawner
\item number of spawning intervals
\item adult carrying capacity
\item sex ratio
}
}
