% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModel_Run.R
\name{PopulationModel_Run}
\alias{PopulationModel_Run}
\title{Population Model Run}
\usage{
PopulationModel_Run(
  dose = NA,
  sr_wb_dat = NA,
  life_cycle_params = NA,
  HUC_ID = NA,
  n_years = 100,
  MC_sims = 10,
  stressors = NA,
  output_type = "full",
  habitat_dd_k = NULL
)
}
\arguments{
\item{dose}{dataframe. Stressor magnitude dataset
imported from StressorMagnitudeWorkbook().}

\item{sr_wb_dat}{list object. Stressor response workbook
imported from StressorResponseWorkbook().}

\item{life_cycle_params}{dataframe. Life cycle parameters.}

\item{HUC_ID}{character. HUC_ID for the location unit. Can only be one HUC_ID (not an array).}

\item{n_years}{numeric. Number of years to run the population.}

\item{MC_sims}{numeric. set number of Monte
Carlo simulations for the Population Model.}

\item{stressors}{(optional) character vector of stressor
names to include in the Population Model. Leave the default
value as NA if you wish to include all stressors
applicable to the population model.}

\item{output_type}{(optional) character. Set to "full" for all data of "adults" for only adult data.}

\item{habitat_dd_k}{(optional) dataframe of location and stage-specific habitat capacity k values for the target species. If used this dataframe will override the capacity estimates}
}
\value{
List object of population model outputs. These include the population vectors for each Monte Carlo simulation and year. List objects names include "ce" for the runs with stressors and "baseline" for the runs without stressors.
}
\description{
Runs the population model with stressors and stressor-response relationships
}
\details{
Runs the population model within
the Joe Model stress-response framework. The population model is
run independently for each single sub-watershed unit (HUC_ID). Inputs for
\code{PopulationModel_Run} include the stressor-response and stressor magnitude
workbooks as well as the life cycle parameters object.
}
\examples{
\dontrun{
library(CEMPRA)
}

}
