% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_DD.R
\name{Projection_DD}
\alias{Projection_DD}
\title{Project matrix with Density Dependence}
\usage{
Projection_DD(
  M.mx = NA,
  D.mx = NULL,
  H.mx = NULL,
  dat = NA,
  Nyears = 100,
  K = NA,
  p.cat = NA,
  CE_df = NULL,
  K_adj = FALSE,
  stage_k_override = NULL,
  bh_dd_stages = NULL,
  anadromous = FALSE,
  alt_return = NULL
)
}
\arguments{
\item{M.mx}{A projection matrix expression}

\item{D.mx}{A matrix of density-dependence effect}

\item{H.mx}{A harm projection matrix}

\item{dat}{Life history data}

\item{Nyears}{Years to run simulation}

\item{K}{The population carrying Capacity of adults (mature individuals)}

\item{p.cat}{Probability of catastrophic event.}

\item{CE_df}{Cumulative effect data frame. Data frame identifying cumulative effects stressors targets system capacity or population parameter, or both, and target life stages.}

\item{K_adj}{Boolean. Should K_adj be run. Defaults to false.}

\item{stage_k_override}{Vector of K values for egg fry (age-0), fry (age-0) to stage_1, stage_1 to stage_2, etc. Defaults to NULL for compensation ratios. If set, values will override adult K value for density dependence with Beverton-Holt. NA values should be used in stages with no constraints. For example, if a fry to stage_1 bottle neck was set to 3000, the resulting vector should look like this c(NA, 3000, NA):
\itemize{
\item 1. Egg to fry (K), NA - no constraint
\item 2. Fry to stage_1 (K), constrained to 3000 fry entering stage_1
\item 3. stage_1 to stage_2 (K), NA - no constraint
}}

\item{bh_dd_stages}{Optional Character vector of life stages c("dd_hs_0", "bh_stage_1", "bh_stage_2", "bh_stage_3", ...) to apply classical Beverton-Holt density-dependence. To be used in place of compensation ratios if set. Use "dd_hs_0" for egg-to-fry k, "bh_stage_1" for fry to stage_1 k and "bh_stage_2" for stage_1 to stage_2 k etc. Densities are the capped value for the transition stage.}

\item{anadromous}{Boolean. If true, the model will apply anadromous life history parameters. Defaults to false.}

\item{alt_return}{Character. Alternative return objects for internal use.}
}
\value{
A list object with projected years, population size, lambda, fecundity, survival, catastrophic events.
}
\description{
Project the matrix model forward in time with density dependence.
}
\details{
The function runs the population projections forward through time.
Life-cycle specific stressors (if set) will be applied based on values in the \code{CE_df} table.
There are several ways to implement density dependence.
deterministic projection matrix using \code{popbio::stable.stage} with initial parameters based on arguments provided. Applies CE stressors to appropriate targets based on \code{CE_df}. All population modeling components are contained within this function. Users define a projection matrix, density-dependence matrix, harm projection matrix, life history parameters, years to simulate, carrying capacity, catastrophic event probabilities and a cumulative effects data frame (CE_df). When run this function will project the population forward in time. See the vignette tutorial Population Model Overview for details and instructions.
}
