% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_LTRE.R
\name{rtruncnorm}
\alias{rtruncnorm}
\title{Sample from Truncated Normal Distribution}
\usage{
rtruncnorm(n, mean, sd, lower, upper)
}
\arguments{
\item{n}{Integer. Number of samples to draw.}

\item{mean}{Numeric. Mean of the underlying normal distribution.}

\item{sd}{Numeric. Standard deviation of the underlying normal distribution.}

\item{lower}{Numeric. Lower truncation bound.}

\item{upper}{Numeric. Upper truncation bound.}
}
\value{
Numeric vector of n samples from the truncated normal distribution.
}
\description{
Sample from a normal distribution truncated to specified bounds
using the inverse CDF method.
}
\details{
Uses the inverse CDF method: samples uniform values between
pnorm(lower) and pnorm(upper), then applies qnorm to obtain truncated
normal samples.
}
\keyword{internal}
