% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{s0_optim_anadromous.f}
\alias{s0_optim_anadromous.f}
\title{s0_optim_anadromous.f}
\usage{
s0_optim_anadromous.f(s0, mx, dat, target.lambda)
}
\description{
s0_optim_anadromous.f
}
\details{
Optimization function - find YOY survival that gives a lambda
}
\keyword{internal}
