% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocioEconomicRun.R
\name{SocioEconomicRun}
\alias{SocioEconomicRun}
\title{Run SocioEconomic Cost Calculations}
\usage{
SocioEconomicRun(socioeconomic_inputs = NA)
}
\arguments{
\item{socioeconomic_inputs}{list object of socioeconomic inputs returned from SocioEconomicWorkbook().}
}
\value{
A list object with cost and area summary calculations included
}
\description{
Provides draft cost calculations for socioeconomic inputs.
}
\details{
Internal function run within the Joe Model to perform socioeconomic calculations.
}
\examples{
\dontrun{
library(CEMPRA)
filename <- system.file("extdata", "socio_economic/socio_economic_input.xlsx", package = "CEMPRA")
socioeconomic_inputs <- SocioEconomicWorkbook(filename = filename)
socioeconomic_inputs <- SocioEconomicRun(socioeconomic_inputs = socioeconomic_inputs)

}

}
