% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocioEconomicWorkbook.R
\name{SocioEconomicWorkbook}
\alias{SocioEconomicWorkbook}
\title{Get Socio Economic Data from Excel Workbook}
\usage{
SocioEconomicWorkbook(filename = NA)
}
\arguments{
\item{filename}{A string. Relative file name to the stressor response Excel workbook.}
}
\value{
A list object of each worksheet organized as separate data frames.
}
\description{
Extract the socio economic data from each sheet
in the Excel workbook.
}
\details{
For more information about data format requirements, see manual.
}
\examples{
\dontrun{
library(CEMPRA)
filename <- system.file("extdata", "socio_economic/socio_economic_input.xlsx", package = "CEMPRA")
socioeconomic_inputs <- SocioEconomicWorkbook(filename = filename)
}

}
