% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_k_for_proj_dd.R
\name{build_k_for_proj_dd}
\alias{build_k_for_proj_dd}
\title{build_k_for_proj_dd}
\usage{
build_k_for_proj_dd(habitat_dd_k, HUC_ID, life_histories, life_cycle_params)
}
\arguments{
\item{life_histories}{Object returned from pop_model_matrix_elements()$life_histories}

\item{life_cycle_params}{Object returned from pop_model_dat_clean() without any further editing}
}
\value{
A list with the following components:
\itemize{
\item ret_stage_k_override: A named numeric vector of carrying capacity (K) values
for each life stage, including fry (K0) and spawners if applicable.
\item ret_bh_dd_stages: A character vector of life stage names that are subject to
Beverton-Holt density dependence.
\item ret_total_anadromous_spawners: A numeric value representing the
total carrying capacity for anadromous spawners, if applicable.
}
}
\description{
Build build_habitat_dd_k object and bh_dd_stages from PopulationModel_Run for Projection_DD.
}
\details{
This function builds the habitat carrying capacity (K) vector for
use in density dependent projections. It extracts K values from the
habitat_dd_k input data frame for the target HUC_ID and formats them for
use in Projection_DD(). It also identifies which life stages are subject
to density dependence based on the life cycle parameters. The function will
return an empty list of there is no location-based DD.
}
